[davinci] Added lots of debug messages to 2D acceleration, set StretchBlit() hook statically.
As there's no other implementation than for 32 bit at the moment, assign the function pointer
for StretchBlit once during driver initialization to save (very minimal) overhead.

+++ gfxdrivers/davinci/davinci_gfxdriver.h
@@ -59,7 +59,7 @@ typedef struct {
      DFBSurfacePixelFormat     src_format;
      unsigned long             src_bpp;
 
-     u32                       source_mult;
+     unsigned long             source_mult;
 
      unsigned long             fillcolor;
 
@@ -67,8 +67,8 @@ typedef struct {
      int                       draw_blend_sub_function;
 
      DFBColor                  color;
-     u32                       color_argb;
-     u32                       colorkey;
+     unsigned long             color_argb;
+     unsigned long             colorkey;
 
      DFBSurfaceBlittingFlags   blitting_flags;
